requirements.txt
absl-py==0.6.1
aiodns==1.1.1
aiohttp==3.5.2
aiohttp-socks==0.2.2
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrs==18.2.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
blinker==1.4
cachetools==3.0.0
cchardet==2.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
docopt==0.6.2
elasticsearch==6.3.1
Flask==1.0.2
Flask-API==1.1
Flask-Bcrypt==0.7.1
Flask-Executor==0.8.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
grep==0.3.2
grpcio==1.16.1
h5py==2.8.0
httplib2==0.12.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
multidict==4.5.2
mysql==0.0.2
mysql-connector-python==8.0.13
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.14
nltk==3.4
numpy==1.15.4
oauth2client==4.1.3
pandas==0.23.4
passlib==1.7.1
Pillow==5.4.1
pipreqs==0.4.9
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycares==2.4.0
pycparser==2.19
PyDrive==1.3.1
PyMySQL==0.9.3
PySocks==1.6.8
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.6.2
SQLAlchemy==1.2.17
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
textblob==0.15.2
textgenrnn==1.4.1
twint==1.2.0.0
typing-extensions==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
volatility==2.6.1
Werkzeug==0.14.1
WTForms==2.2.1
yarg==0.1.9
yarl==1.3.0
